backend/requirements/base.txt
# ── Core ──────────────────────────────────────────────────────────
Django==5.0.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# ── Database ──────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── Caching / Async / WebSockets ──────────────────────────────────
redis==5.0.8
django-redis==5.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# ── Celery ────────────────────────────────────────────────────────
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1

# ── CORS ──────────────────────────────────────────────────────────
django-cors-headers==4.4.0

# ── Filtering / Search ────────────────────────────────────────────
django-filter==24.3

# ── Phone numbers ─────────────────────────────────────────────────
phonenumbers==8.13.44
django-phonenumber-field==7.3.0

# ── File/Image handling ───────────────────────────────────────────
Pillow==10.4.0

# ── Environment config ────────────────────────────────────────────
python-decouple==3.8

# ── API Documentation ─────────────────────────────────────────────
drf-spectacular==0.27.2

# ── HTTP client (payment provider calls) ─────────────────────────
httpx==0.27.2

# ── Utilities ────────────────────────────────────────────────────
python-dateutil==2.9.0
pytz==2024.2
shortuuid==1.0.13
